Try to use zuul-cloner to prepare fixtures
In OpenStack Infra, we would like to run Puppet unit tests that
sometimes depends on other Puppet OpenStack modules.
Example: a patch in puppet-openstacklib that needs to be tested in puppet-nova.

This patch modifies the Rakefile to:
* clean spec_prep and spec_clean Rake tasks
* use openstack/puppet-openstack-integration/install_modules_unit.sh script
  to clone modules.
* do not use .fixtures.yaml file to clone modules and rely on
  zuul-cloner or git.
* Add openstack/ in gitignore so we never commit the
  puppet-openstack-integration repository (can happen when spec_clean
  did not run but you want to submit the patch anyway)
* Allow to run a custom Puppetfile if PUPPETFILE env is exported. It
  will allow people to test the module with the dependencies they like,
  feature we had with .fixtures.yaml.

Also add 'r10k' to Gemfile.

That way, we will be able to use zuul dependencies and run tests accross
modules like we do with functional testing.

Change-Id: I8b884ef477bbc031da6bd3955ed308679e0380f6

release: prepare 7.0.0 (liberty)

This patch prepares this Puppet module to be released in 7.x series (Liberty)
with 7.0.0.

Change-Id: I4d892d41a8d7c8ed1f1a00712afced944c821f92

Update rspec tests for keystone

https://review.openstack.org/#/c/228744/ has been merged, we need to
drop the 'tenant' parameter to Keystone_user resource.

Change-Id: I515769d73b26b21b2a18914411a9d9aee27f95c9